requirements.txt
aiohttp==3.6.2
alpaca-trade-api==0.38
alpha-vantage==2.1.0
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.2.5
async-timeout==3.0.1
asyncio-nats-client==0.9.2
attrs==19.3.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
cycler==0.10.0
discord.py==1.3.3
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.8
isort==4.3.20
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
matplotlib==3.1.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.17.1
pandas==0.25.1
parse==1.12.0
prettytable==0.7.2
psycopg2==2.8.5
pyee==6.0.0
pylint==2.3.1
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2019.2
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.1
texttable==1.6.2
tqdm==4.32.1
typed-ast==1.4.0
urllib3==1.24.3
w3lib==1.20.0
websocket-client==0.56.0
websockets==8.0.2
whitenoise==5.1.0
wrapt==1.11.1
yarl==1.4.2